[Refactor] #111 - Move cartDiffUtil
