+++ presentation/src/main/java/com/woowahan/banchan/ui/adapter/DefaultCartAdapter.kt
@@ -1,13 +1,9 @@
 package com.woowahan.banchan.ui.adapter
 
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import androidx.core.widget.doAfterTextChanged
-import androidx.core.widget.doOnTextChanged
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
-import com.woowahan.banchan.CartListModelDiffUtilCallback
 import com.woowahan.banchan.databinding.ItemCartContentBinding
 import com.woowahan.banchan.databinding.ItemCartFooterBinding
 import com.woowahan.banchan.databinding.ItemCartHeaderBinding
@@ -271,5 +267,49 @@ class DefaultCartAdapter(
         object SelectOneChanged : Payload()
         object quantityChanged : Payload()
         object totalPriceChanged : Payload()
+    }
+
+    class CartListModelDiffUtilCallback(
+        private val oldList: List<CartListItemModel>,
+        private val newList: List<CartListItemModel>
+    ) : DiffUtil.Callback() {
+        override fun getOldListSize(): Int = oldList.size
+
+        override fun getNewListSize(): Int = newList.size
+
+        override fun areItemsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
+            return oldList[oldItemPosition] isSameIdWith newList[newItemPosition]
+        }
+
+        override fun areContentsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
+            return oldList[oldItemPosition] isSameContentWith newList[newItemPosition]
+        }
+
+        override fun getChangePayload(oldItemPosition: Int, newItemPosition: Int): Any? {
+            val oldItem = oldList[oldItemPosition]
+            val newItem = newList[newItemPosition]
+            when {
+                oldItem is CartListItemModel.Header && newItem is CartListItemModel.Header -> {
+                    if (!(oldItem isSameContentWith newItem)) {
+                        return DefaultCartAdapter.Payload.SelectAllChanged
+                    }
+                }
+                oldItem is CartListItemModel.Content && newItem is CartListItemModel.Content -> {
+                    if (oldItem.cart.isSelected != newItem.cart.isSelected) {
+                        return DefaultCartAdapter.Payload.SelectOneChanged
+                    }
+                    if (oldItem.cart.count != newItem.cart.count) {
+                        return DefaultCartAdapter.Payload.quantityChanged
+                    }
+                }
+                oldItem is CartListItemModel.Footer && newItem is CartListItemModel.Footer -> {
+                    if (!(oldItem isSameContentWith newItem)) {
+                        return DefaultCartAdapter.Payload.totalPriceChanged
+                    }
+                }
+            }
+
+            return super.getChangePayload(oldItemPosition, newItemPosition)
+        }
     }
 }
